% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capushe.R
\docType{class}
\name{Capushe-class}
\alias{Capushe-class}
\title{Class "Capushe"}
\arguments{
\item{object}{an object with class \code{capushe}}
}
\description{
Class of object returned by the \code{capushe} function.
}
\section{Slots}{

\describe{
\item{\code{DDSE}}{A list returned by the \code{\link[=DDSE]{DDSE}} function.}

\item{\code{Djump}}{A list returned by the \code{\link[=Djump]{Djump}} function.}

\item{\code{AIC_capushe}}{A list returned by the \code{\link[=AICcapushe]{AICcapushe}} function.}

\item{\code{BIC_capushe}}{A list returned by the \code{\link[=BICcapushe]{BICcapushe}} function.}

\item{\code{n}}{The number of observations given by the user.}
}}

\references{
Article: Baudry, J.-P., Maugis, C. and Michel, B. (2011) Slope heuristics:
overview and implementation. \var{Statistics and Computing}, to appear. doi: 10.1007/s11222-011-9236-1
}
\seealso{
See also \code{\link{plot,Capushe-method}} and \code{\link{capushe}}.
}
